Regex
In this assignment, you will work on regular expression. For simplicity, we will assume that
there is a fixed set of regular expressions. We will not consider out of these. You must use any
built-in method or package in your implementation. The following table contains a fixed set of
RE that will be used in this assignment.